fix: modified code to move auth related pages & compo. in common pckg
- moved sign-in page & its components
- moved sign-up page & its components
- this compo. are configurable & can be used across apps
