+++ apps/retail/pages/signUp.tsx
@@ -1,171 +1,13 @@
-import React, { useMemo, useState } from 'react'
-import { useLanguage } from '@hooks/useLanguage'
-import AlternateLogo from '../public/images/KuzaLogo.svg'
-import { signUpValidateForm } from '@beckn-ui/common/src/utils'
-import { BecknAuth } from '@beckn-ui/becknified-components'
-import Router from 'next/router'
-import { Box, useBreakpoint } from '@chakra-ui/react'
-import Logo from '@public/images/Logo.svg'
-import { FetchBaseQueryError } from '@reduxjs/toolkit/query'
-import { FormErrors, SignInResponse, SignUpProps } from '@beckn-ui/common/lib/types'
-import { useRegisterMutation } from '@beckn-ui/common/src/services/User'
+import { Box } from '@chakra-ui/react'
+import React from 'react'
+import SignUp from '@components/signUp/SignUp'
 
-const SignUp = () => {
-  const { t } = useLanguage()
-
-  const [formData, setFormData] = useState<SignUpProps>({ name: '', email: '', password: '', mobileNumber: '' })
-  const [formErrors, setFormErrors] = useState<FormErrors>({ name: '', email: '', password: '', mobileNumber: '' })
-  const breakpoint = useBreakpoint()
-  const [register, { isLoading, isError }] = useRegisterMutation()
-
-  const baseUrl = process.env.NEXT_PUBLIC_STRAPI_URL
-  const mobileBreakpoints = ['base', 'sm', 'md', 'lg']
-  const currentLogo = mobileBreakpoints.includes(breakpoint) ? Logo : AlternateLogo
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target
-
-    setFormData((prevFormData: SignUpProps) => ({
-      ...prevFormData,
-      [name]: value
-    }))
-
-    const updatedFormData = {
-      ...formData,
-      [name]: value
-    }
-
-    const errors = signUpValidateForm(updatedFormData)
-    setFormErrors(prevErrors => ({
-      ...prevErrors,
-      [name]: t[`${errors[name as keyof FormErrors]}`] || ''
-    }))
-  }
-
-  const handleSignUp = async () => {
-    const errors = signUpValidateForm(formData)
-    const isFormValid = Object.values(errors).every(error => error === '')
-
-    if (isFormValid) {
-      try {
-        const registerResponse = await register({
-          username: formData.email,
-          email: formData.email,
-          password: formData.password,
-          mobile: formData.mobileNumber
-        })
-
-        if ((registerResponse as { error: FetchBaseQueryError })?.error) throw new Error('Could not register')
-
-        const myHeaders = new Headers()
-        myHeaders.append('Authorization', `Bearer ${(registerResponse as { data: SignInResponse }).data.jwt}`)
-
-        const currentFormData = new FormData()
-        const data = {
-          name: formData.name,
-          phone: formData.mobileNumber
-        }
-
-        currentFormData.append('data', JSON.stringify(data))
-
-        const requestOptions: RequestInit = {
-          method: 'POST',
-          headers: myHeaders,
-          redirect: 'follow',
-          body: currentFormData
-        }
-
-        fetch(`${baseUrl}/profiles`, requestOptions).then(response => {
-          Router.push('/')
-          return response.json()
-        })
-      } catch (error) {
-        console.error('An error occurred:', error)
-      }
-    } else {
-      setFormErrors({
-        ...formErrors,
-        name: t[`${errors.name}`] || ''
-      })
-    }
-  }
-
-  const isFormFilled = useMemo(() => {
-    return (
-      Object.values(formData).every(value => value !== '') && Object.values(formErrors).every(value => value === '')
-    )
-  }, [formData, formErrors])
-
+const Register = () => {
   return (
-    <Box
-      mt={'30px'}
-      className="hideScroll"
-      maxH="calc(100vh - 90px)"
-      overflowY={'scroll'}
-    >
-      <BecknAuth
-        schema={{
-          logo: {
-            src: currentLogo,
-            alt: 'Suppliflow logo'
-          },
-          buttons: [
-            {
-              text: t.signUp,
-              handleClick: handleSignUp,
-              disabled: !isFormFilled,
-              variant: 'solid',
-              colorScheme: 'primary',
-              isLoading: isLoading
-            },
-            {
-              text: t.signIn,
-              handleClick: () => {
-                Router.push('/signin')
-              },
-              variant: 'outline',
-              colorScheme: 'primary',
-              disabled: isLoading
-            }
-          ],
-          inputs: [
-            {
-              type: 'text',
-              name: 'name',
-              value: formData.name,
-              handleChange: handleInputChange,
-              label: t.fullName,
-              error: formErrors.name
-            },
-            {
-              type: 'text',
-              name: 'email',
-              value: formData.email,
-              handleChange: handleInputChange,
-              label: t.enterEmailID,
-              error: formErrors.email
-            },
-            {
-              type: 'password',
-              name: 'password',
-              value: formData.password,
-              handleChange: handleInputChange,
-              label: t.enterPassword,
-              error: formErrors.password
-            },
-            {
-              type: 'number',
-              name: 'mobileNumber',
-              value: formData.mobileNumber,
-              handleChange: handleInputChange,
-              label: t.enterMobileNumber,
-              error: formErrors.mobileNumber
-            }
-          ]
-        }}
-      />
+    <Box>
+      <SignUp />
     </Box>
   )
 }
 
-export default SignUp
+export default Register
