+++ packages/common/index.ts
@@ -4,3 +4,4 @@ export * from './lib/config'
 export * from './src/services'
 export * from './src/hooks'
 export * from './src/utils'
+export * from './src/pages'
