+++ packages/common/src/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as SignInPage } from './signin'
+export { default as SignUpPage } from './signup'
